requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
PySocks==1.7.1
stem==1.8.1
fake-useragent==1.4.0
argparse==1.4.0
jinja2==3.1.3
html5lib==1.1
lxml==4.9.3 